requirements.txt
fastapi>=0.104.0
uvicorn>=0.24.0
sqlalchemy>=2.0.23
alembic>=1.12.1
psycopg2-binary>=2.9.9
python-dotenv>=1.0.0
pydantic>=2.4.2
pytest>=7.4.3
httpx>=0.25.1
python-multipart>=0.0.6
black>=23.10.1
isort>=5.12.0
email-validator>=2.1.0
